+++ player/player.ts
@@ -1,6 +1,6 @@
 import videoJs, { VideoJsPlayerOptions } from 'video.js'
-import { props as _props, Props, PropKey } from './props'
-import { eventKeys, EventKey } from './events'
+import { propsConfig, Props, PropKey } from './props'
+import { events, EventKey } from './events'
 import type { VideoJsPlayer } from './type'
 
 export interface CreatePlayerOptions {
@@ -11,10 +11,10 @@ export interface CreatePlayerOptions {
 
 export type PlayerResult = ReturnType<typeof createPlayer>
 export const createPlayer = ({ props, element, onEvent }: CreatePlayerOptions) => {
-  // exclude fallback options
+  // Exclude fallback options.
   const { options: fallbackOptions = {}, ...optProps } = props
 
-  // exclude undefined value
+  // Exclude undefined value.
   const propOptions: Omit<Props, 'options'> = {}
   const optPropKeys = Object.keys(optProps) as Array<keyof typeof optProps>
   optPropKeys.forEach((key) => {
@@ -25,7 +25,7 @@ export const createPlayer = ({ props, element, onEvent }: CreatePlayerOptions) =
     }
   })
 
-  // merge fallback options & exclude component options
+  // Merge fallback options & exclude component options.
   const { volume, playbackRate, ...initOptions } = {
     ...propOptions,
     ...fallbackOptions
@@ -33,10 +33,10 @@ export const createPlayer = ({ props, element, onEvent }: CreatePlayerOptions) =
 
   // init player
   const player = videoJs(element, initOptions, function () {
-    // stringing video.js events to vue emits
-    eventKeys.forEach((eventKey) => {
-      this.on(eventKey, (events) => {
-        onEvent(eventKey, events)
+    // Stringing video.js events to vue emits.
+    events.forEach((eventKey) => {
+      this.on(eventKey, (payload) => {
+        onEvent(eventKey, payload)
       })
     })
 
@@ -57,36 +57,30 @@ export const createPlayer = ({ props, element, onEvent }: CreatePlayerOptions) =
     // https://github.com/sampotts/plyr/blob/master/src/js/plyr.js#L677
     // https://github.com/bytedance/xgplayer/blob/master/packages/xgplayer/src/skin/controls/playbackRate.js#L30
     if (playbackRate && Number.isFinite(playbackRate)) {
-      // video always reads defaultPlaybackRate as the initial playbackRate when switching video sources.
+      // Video always reads defaultPlaybackRate as the initial playbackRate when switching video sources.
       this.defaultPlaybackRate(playbackRate)
-      // ensures that all synchronization code has been executed by the time playbackRate is executed.
+      // Ensures that all synchronization code has been executed by the time playbackRate is executed.
       setTimeout(() => {
         this.playbackRate(playbackRate)
       }, 0)
     }
   }) as VideoJsPlayer
 
-  // set new class names to Video.js container element
+  // Set new class names to Video.js container element.
   const updateClassNames = (oldClassName: string | void, newClassName: string | void) => {
-    const oNames = oldClassName?.split(' ')
-    if (oNames?.length) {
-      oNames.map((name) => player.removeClass(name))
-    }
-    const nNames = newClassName?.split(' ')
-    if (nNames?.length) {
-      nNames.map((name) => player.addClass(name))
-    }
+    oldClassName?.split(' ').map((name) => player.removeClass(name))
+    newClassName?.split(' ').map((name) => player.addClass(name))
   }
 
-  // set new options to Video.js
+  // Set new options to Video.js config.
   const updateOptions = (options: VideoJsPlayerOptions) => {
     player.options?.(options ?? {})
   }
 
-  // set new prop value to Video.js config
+  // Set new prop value to Video.js config.
   const updatePropOption = <K extends PropKey>(key: K, value: Props[K]) => {
     updateOptions({ [key]: value })
-    _props[key]?.onChange?.(player, value as never)
+    propsConfig[key]?.onChange?.(player, value as never)
   }
 
   return {