+++ packages/react/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+### v0.2.0 (2022-08-05)
+
+- fix: state `playbackRate`
+- feat: `state.fullscreen` > `state.isFullscreen`
+- feat: add new state `isInPictureInPicture`
+- feat: add new state `isLive`
+- feat: new prop `disablePictureInPicture`
+- feat: new [div ingest](https://videojs.com/guides/embeds/#player-div-ingest) type
+
 ### v0.1.0 (2022-07-28)
 
 - feat: full support for the Video.js API